[FIX] sale_stock: move never executed tests
Commit [1] adds some tests on Odoo 14.0 in a file named
`test_report`. However, from 15.0 on, the file has became
`test_sale_stock_report` but the tests of [1] have not been moved in
that new file . Instead, a new file `test_report` has been created
and did not have been added in `__init__` (see [2]). Therefore, the
tests are never executed

[1] 012e25f27ffc0bc51682c06896c9116944661011
[2] 15194cec233332e787b9f546e39c6127301ff9aa
